Pushing a fix for auth and homepage refresh
